requirements.txt
pandas~=2.2.3
opencv-python~=4.11.0.86
numpy~=2.2.5
pillow~=10.4.0
supervision~=0.25.1
tqdm~=4.67.1
torch~=2.7.0+cu126
parse~=1.20.2
ultralytics~=8.3.122
torchvision~=0.22.0+cu126
dash~=3.0.4
plotly~=6.0.1
matplotlib~=3.10.1
seaborn~=0.13.2
ipython~=8.32.0
fastapi~=0.115.12
uvicorn~=0.34.2
cloudinary~=1.44.0